Refactor assembly compiler for web: Introduce common base implementation
- Moved common assembly compilation logic to `creatorCompilerBase.mjs` for reuse between web and Deno.
- Updated `assembly_compiler_default` to utilize the new base implementation.
- Enhanced error handling with formatted error messages, now all using ANSI escape codes.
- Adjusted imports in `ToolbarBtngroup.vue` to reflect changes in the assembly compiler structure.
